Fixed bug with type conversion

+++ include/window.hpp
@@ -27,10 +27,10 @@ private:
     float windowWidth;
     float aspectRatio;
 
-    float delta;
-    float totalTime;
-    float frameTime;
-    float lastTime;
+    double delta;
+    double totalTime;
+    double frameTime;
+    double lastTime;
 
     GLFWwindow* window;
